golem_base_mock::api
Trait GolemBaseRpcServer

pub trait GolemBaseRpcServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn get_entity<'life0, 'async_trait>(
        &'life0 self,
        key: B256,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Entity>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_entity_metadata<'life0, 'async_trait>(
        &'life0 self,
        key: B256,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<EntityMetaData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_entity_count<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_all_entity_keys<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<B256>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_entities_of_owner<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<B256>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_storage_value<'life0, 'async_trait>(
        &'life0 self,
        keys: B256,
    ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_entities<'life0, 'async_trait>(
        &'life0 self,
        query: String,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SearchResult>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_entities_to_expire_at_block<'life0, 'async_trait>(
        &'life0 self,
        block_number: u64,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<B256>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the GolemBaseRpc RPC API.

Required Methods

fn get_entity<'life0, 'async_trait>( &'life0 self, key: B256, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Entity>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_entity_metadata<'life0, 'async_trait>( &'life0 self, key: B256, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<EntityMetaData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_entity_count<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_all_entity_keys<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<B256>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_entities_of_owner<'life0, 'async_trait>( &'life0 self, address: Address, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<B256>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_storage_value<'life0, 'async_trait>( &'life0 self, keys: B256, ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn query_entities<'life0, 'async_trait>( &'life0 self, query: String, ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SearchResult>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_entities_to_expire_at_block<'life0, 'async_trait>( &'life0 self, block_number: u64, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<B256>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.