golem_base_mock
Module controller

Structs

EndpointCallback

Wrapper object that provides an await function to wait for endpoint trigger

MockController

Controller for managing mock market responses

WithCallback

Generic wrapper for any response type that can include notification. It sends a notification on drop.

Enums

CallOverride

CallResponse

Response types that can be forced for RPC calls.

CallbackResult

Result of waiting for an endpoint to be triggered